fix: Moved validate_query_rnaseqGeneCount function from mds3.init.js to termdb.DE.ts. Fixed related typing errors caused by moving from javascript to typescript. Updated documentation of validate_query_rnaseqGeneCounts in the style of JSDoc for easier understanding in the future.
